refactor(proxy): extract auth-gate module
Move SessionInfo, sessionCache, getSessionInfo, setSessionCache,
extractSessionToken, and SESSION_CACHE_TTL_MS from src/proxy.ts to
src/lib/proxy/auth-gate.ts. Also export hasSessionCookie helper that
the upcoming csrf-gate (C4) will use for request-attribute-based
gating.

src/proxy.ts re-exports as test-only shims (_sessionCache,
_setSessionCache, _extractSessionToken) so src/__tests__/proxy.test.ts
continues to pass without modification.

This is C2 of the proxy-csrf-enforcement plan.
